Bug 1153324 - Submit to OrangeFactor's API rather than ES directly
Since Elasticsearch is protected by VPN (ES doesn't do auth) and setting
up access is more hassle than it's worth on Heroku. Instead a new API
endpoint has been added to OrangeFactor that forwards our classification
report on for us.

A hardcoded API key is used for authentication, which will need to be
set in the environment of prod before this lands.